requirements.txt
transformers
torch
Pillow
requests
flask
